wareflow_p1/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
xgboost==2.1.0
joblib==1.4.2
firebase-admin==6.5.0
numpy==1.26.4
pandas==2.2.2
python-dotenv==1.0.1
httpx==0.27.0
scikit-learn==1.5.1
